+++ back/prisma/migrations/147_create_bsff_recepisse_exemption.sql
@@ -1,0 +1,6 @@
+ALTER TABLE
+    "default$default"."Bsff"
+ADD COLUMN
+    "transporterRecepisseIsExempted" BOOLEAN DEFAULT FALSE;
+
+UPDATE "default$default"."Bsff" SET "transporterRecepisseIsExempted" = true WHERE "transporterRecepisseNumber" IS NULL;