+++ back/src/bsffs/recipify.ts
@@ -8,9 +8,6 @@ import { recipifyGeneric } from "../companies/recipify";
 const bsffAccessors = (input: BsffInput) => [
   {
     getter: () =>
-      // Old way of exempting is null
-      input?.transporter?.recepisse !== null &&
-      // New way added to the old
       input?.transporter?.recepisse?.isExempted !== true
         ? input?.transporter?.company
         : null,