+++ front/src/Apps/common/queries/fragments/bsdd.ts
@@ -447,6 +447,7 @@ export const dashboardFormFragment = gql`
       company {
         siret
         orgId
+        name
       }
       numberPlate
       customInfo